// src/app/types/address/address.ts
export type AddressSearchCEP = {
  city: string;
  complement: string;
  state: string;
  street: string;
  neighborhood: string;
};

export type AddressUser = {
  city: string;
  complement: string;
  state: string;
  street: string;
  district: string;
  phoneNumber: string;
  addressNumber: number;
  addressCodePostal: string;
};